fix(visualizations): refactor legend set handling and improve visualization components
- Consolidated legend set retrieval into a shared utility method.
- Streamlined visualization components by removing redundant legend set logic.
- Enhanced i18n strings and fixed mismatched keys for better localization support.
- Updated API query fields to include legend-related details for improved visualization context.
