+++ apps/portal/src/components/displayItems/visualizations/DataVisualization.tsx
@@ -2,54 +2,18 @@ import { ReadonlyURLSearchParams } from "next/navigation";
 import { dhis2HttpClient } from "@/utils/api/dhis2";
 import {
 	ChartVisualizationItem,
-	LegendSetConfig,
 	VisualizationChartType,
 	VisualizationConfig,
 } from "@packages/shared/schemas";
 import { getAppearanceConfig } from "@/utils/config/appConfig";
 import { YearOverYearDataVisComponent } from "@/components/displayItems/visualizations/YearOverYearDataVisComponent";
 import { DataVisComponent } from "@/components/displayItems/visualizations/DataVisComponent";
-import { LegendSet } from "@hisptz/dhis2-utils";
-import { camelCase, compact, get } from "lodash";
 
 export interface MainVisualizationProps {
 	searchParams?: ReadonlyURLSearchParams;
 	config: ChartVisualizationItem;
 	showFilter?: boolean;
 	disableActions?: boolean;
-}
-
-async function getVisualizationLegendSet(
-	config: VisualizationConfig,
-): Promise<LegendSetConfig | undefined> {
-	if (!config.legend) {
-		return;
-	}
-	const legendConfig = config.legend;
-	switch (legendConfig.strategy) {
-		case "FIXED":
-			if (!legendConfig.set?.id) {
-				return;
-			}
-			return legendConfig.set as LegendSet;
-		case "BY_DATA_ITEM":
-			return compact(
-				config.dataDimensionItems.map((item) => {
-					const dataItem = get(
-						item,
-						camelCase(item.dataDimensionItemType.toLowerCase()),
-					);
-
-					if (dataItem.legendSet) {
-						return {
-							dataItem: dataItem.id,
-							legendSet: dataItem.legendSet as LegendSet,
-						};
-					}
-					return undefined;
-				}),
-			);
-	}
 }
 
 export async function getDataVisualization(config: ChartVisualizationItem) {
@@ -65,7 +29,6 @@ export async function getDataVisualization(config: ChartVisualizationItem) {
 
 	return {
 		visualizationConfig,
-		legendSetConfig: await getVisualizationLegendSet(visualizationConfig),
 	};
 }
 
@@ -74,8 +37,7 @@ export async function DataVisualization({
 	disableActions,
 	showFilter,
 }: MainVisualizationProps) {
-	const { visualizationConfig, legendSetConfig } =
-		await getDataVisualization(config);
+	const { visualizationConfig } = await getDataVisualization(config);
 	const { appearanceConfig } = (await getAppearanceConfig())!;
 	const colors = appearanceConfig.colors.chartColors;
 
@@ -105,7 +67,6 @@ export async function DataVisualization({
 
 	return (
 		<DataVisComponent
-			legendSetConfig={legendSetConfig}
 			showFilter={showFilter}
 			colors={colors}
 			disableActions={disableActions}
