+++ packages/ui/src/visualizations/ChartSelector.tsx
@@ -3,7 +3,6 @@ import React, { RefObject } from "react";
 import HighchartsReact from "highcharts-react-official";
 import {
 	AnalyticsData,
-	LegendSetConfig,
 	VisualizationChartType,
 	VisualizationConfig,
 } from "@packages/shared/schemas";
@@ -18,7 +17,6 @@ export function ChartSelector({
 	setRef,
 	fullScreen,
 	colors,
-	legendSetConfig,
 }: {
 	setRef: RefObject<HighchartsReact.RefObject | null>;
 	analytics: AnalyticsData;
@@ -26,7 +24,6 @@ export function ChartSelector({
 	tableRef: RefObject<HTMLTableElement | null>;
 	fullScreen: boolean;
 	colors: string[];
-	legendSetConfig?: LegendSetConfig;
 }) {
 	const chartType = visualization.type;
 	switch (chartType) {
@@ -54,7 +51,6 @@ export function ChartSelector({
 					setRef={setRef}
 					analytics={analytics}
 					visualization={visualization}
-					legendSet={legendSetConfig}
 				/>
 			);
 	}
